improved waste import

+++ app/Console/Commands/ImportWasteStreets.php
@@ -2,11 +2,11 @@
 
 namespace App\Console\Commands;
 
-use App\Imports\WasteDatesImport;
 use App\Imports\WasteStreetsImport;
 use App\Models\RubbishStreet;
-use Excel;
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\DB;
+use Maatwebsite\Excel\Facades\Excel;
 
 class ImportWasteStreets extends Command
 {
@@ -15,25 +15,39 @@ class ImportWasteStreets extends Command
      *
      * @var string
      */
-    protected $signature = 'waste:import-streets';
+    protected $signature = 'waste:import-streets {file=street_register.csv : The CSV file to import}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Imports a csv of a rubbish collection streets';
+    protected $description = '';
 
     /**
      * Execute the console command.
      *
      * @return int
      */
-    public function handle()
+    public function handle(): int
     {
-        RubbishStreet::query()->truncate();
+        $file = $this->argument('file');
 
-        Excel::import(new WasteStreetsImport(), 'street_register_2023.csv');
+        // Check if the file exists
+        if (!file_exists($file)) {
+            $this->error("The file {$file} does not exist.");
+            return Command::FAILURE;
+        }
+
+        DB::transaction(function () use ($file) {
+            // Truncate the table
+            RubbishStreet::query()->truncate();
+
+            // Import the data
+            Excel::import(new WasteStreetsImport(), $file);
+        });
+
+        $this->info("The file {$file} has been successfully imported.");
 
         return Command::SUCCESS;
     }
